serve/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
torch==2.2.0
transformers==4.41.0
azure-storage-blob==12.26.0
numpy==1.26.4